Move item definitions logic into context
